Removendo campos da tela de cadastro e deletando a classe RegisterFooterView()
