+++ Cadastro e Login Firebase/Features/Register/RegisterView.swift
@@ -9,6 +9,23 @@ import UIKit
 
 class RegisterView: UIView {
     
+    var registerAction: (() -> Void)?
+    var backAction: (() -> Void)?
+    
+    static let height: CGFloat = 100
+    
+    private lazy var btRegister: DefaultButton = {
+        let view = DefaultButton(title: "Cadastrar")
+        view.addTarget(self, action: #selector(performRegister), for: .touchUpInside)
+        return view
+    }()
+    
+    private lazy var btBack: DefaultButton = {
+        let view = DefaultButton(title: "Voltar")
+        view.addTarget(self, action: #selector(performBack), for: .touchUpInside)
+        return view
+    }()
+    
     private lazy var tableView: UITableView = {
         //Delegate is register by ViewController for func setupTableViewDelegate
         let view = UITableView(frame: .zero, style: .plain)
@@ -31,6 +48,8 @@ class RegisterView: UIView {
         
     private func setupView() {
         self.addSubview(tableView)
+        self.addSubview(btBack)
+        self.addSubview(btRegister)
     }
     
     private func setupConstraints() {
@@ -38,12 +57,31 @@ class RegisterView: UIView {
             self.tableView.topAnchor.constraint(equalTo: self.topAnchor),
             self.tableView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
             self.tableView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            self.tableView.bottomAnchor.constraint(equalTo: self.bottomAnchor)
+            
+            self.btBack.bottomAnchor.constraint(equalTo: self.safeAreaLayoutGuide.bottomAnchor, constant: -16),
+            self.btBack.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 16),
+            self.btBack.trailingAnchor.constraint(equalTo: self.centerXAnchor, constant: -8),
+            self.btBack.heightAnchor.constraint(equalToConstant: 44),
+            
+            self.btRegister.bottomAnchor.constraint(equalTo: self.btBack.bottomAnchor),
+            self.btRegister.widthAnchor.constraint(equalTo: self.btBack.widthAnchor),
+            self.btRegister.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -16),
+            self.btRegister.heightAnchor.constraint(equalToConstant: 44),
+            
+            self.tableView.bottomAnchor.constraint(equalTo: self.btBack.topAnchor, constant: -16)
         ])
     }
     
     private func setupAditionalConfigurations() {
         self.backgroundColor = .white
+    }
+    
+    @objc private func performRegister() {
+        self.registerAction?()
+    }
+    
+    @objc private func performBack() {
+        self.backAction?()
     }
     
     public func setupTableView() {
