+++ Cadastro e Login Firebase/Resources/AppDelegate.swift
@@ -8,7 +8,6 @@
 import UIKit
 import FirebaseCore
 import FirebaseAuth
-import IQKeyboardManagerSwift
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -17,7 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         FirebaseApp.configure()
-        //IQKeyboardManager.shared.enable = true
         return true
     }
